abm aeronave, abm ruta -- falta eliminar en ambos

+++ src/AerolineaFrba/Models/Ruta.cs
@@ -9,8 +9,8 @@ namespace AerolineaFrba.Models {
 
         public int Id { get; set; }
         public int Codigo { get; set; }
-        public float Precio_Base_Kg { get; set; }
-        public float Precio_Base_Pasajes { get; set; }
+        public decimal Precio_Base_Kg { get; set; }
+        public decimal Precio_Base_Pasajes { get; set; }
 
         public int Ciudad_Origen_Id { get; set; }
         public Ciudad CiudadOrigen { get; set; }
